+++ include/siri/net/promise.h
@@ -42,7 +42,7 @@ typedef struct sirinet_promise_s
     uv_timer_t * timer;
     sirinet_promise_cb cb;
     siridb_server_t * server;
-    uint64_t pid;
+    uint32_t pid;
     void * data;
 } sirinet_promise_t;
